requirements.txt
# Python dependencies
python-telegram-bot==21.0
google-generativeai==0.8.3
python-dotenv==1.0.1
numpy>=1.26.0

# Simple vector storage (alternative to ChromaDB for Python 3.14)
scikit-learn>=1.3.0

# Optional but recommended
fastapi==0.115.6
uvicorn==0.34.0
